Defer running check to a predicate method
We need to have better error handling and returns for our detection of
whether the service is running. This change implements a new method in
the service resource, `is_service_up?` that performs this in a more
robust manner. The method has a comment that describes its behavior in
greater detail.

Fixes #58

Signed-off-by: Joshua Timberman <joshua@chef.io>