Create the hab user's home directory
I was debugging one of my Habitat packages and started running across
applications that don't behave correctly if a user's home directory doesn't
exist. Specifically, the `irb` in Chef's embedded Ruby fails to launch and
Inspec has some config options that default to the home dir.

I'm not sure how common this is but, it feels like, if `/home/hab` is going to
be set as the user's home directory, it'd be appropriate for it to exist.

Signed-off-by: Jonathan Hartman <j@hartman.io>